Add files Via Upload
made a command line program to tell you your daily, weekly, monthly, and yearly compensation. Your results depend on wether you are salaried or hourly and what you are paid either in the year or per hour.  Salary part of the program is not working at the moment but hourly is.